+++ scanapi/utils.py
@@ -1,4 +1,9 @@
+import requests
+from requests.adapters import HTTPAdapter
+from urllib3.util.retry import Retry
+
 from scanapi.errors import InvalidKeyError, MissingMandatoryKeyError
+from scanapi.tree.tree_keys import MAX_RETRIES_KEY
 
 
 def join_urls(first_url, second_url):
@@ -35,3 +40,21 @@ def _validate_required_keys(keys, required_keys, scope):
     if not set(required_keys) <= set(keys):
         missing_keys = set(required_keys) - set(keys)
         raise MissingMandatoryKeyError(missing_keys, scope)
+
+
+def session_with_retry(retry_configuration):
+    """ Instantiate a requests session with the retry configuration if provided by
+    instantiating an HTTPAdapter mounting it into the mounting it into the
+    `requests.Session`.
+    """
+    session = requests.Session()
+
+    if not retry_configuration:
+        return session
+
+    retry = Retry(total=retry_configuration.get(MAX_RETRIES_KEY, 0))
+    adapter = HTTPAdapter(max_retries=retry)
+    session.mount("http://", adapter)
+    session.mount("https://", adapter)
+
+    return session