+++ tests/unit/test_utils.py
@@ -2,7 +2,7 @@ import pytest
 import requests
 
 from scanapi.errors import InvalidKeyError, MissingMandatoryKeyError
-from scanapi.utils import join_urls, validate_keys
+from scanapi.utils import join_urls, session_with_retry, validate_keys
 
 
 @pytest.fixture
@@ -97,3 +97,18 @@ class TestValidateKeys:
             scope = "endpoint"
 
             validate_keys(keys, available_keys, mandatory_keys, scope)
+
+
+class TestSessionWithRetry:
+    class TestNoRetryConfiguration:
+        def test_should_not_mount_custom_adapters(self):
+            session = session_with_retry({})
+
+            assert session.adapters["http://"].max_retries.total == 0
+            assert session.adapters["https://"].max_retries.total == 0
+
+        def test_should_mount_custom_adapters(self):
+            session = session_with_retry({"max_retries": 7})
+
+            assert session.adapters["http://"].max_retries.total == 7
+            assert session.adapters["https://"].max_retries.total == 7